[bufio] Change lru->count to ulong.
In case someone has terabytes of memory to devote to bufio alone.

+++ drivers/md/dm-bufio.c
@@ -82,7 +82,7 @@ struct lru_entry {
 
 struct lru {
 	struct list_head *cursor;
-	unsigned count;
+	unsigned long count;
 };
 
 /*------------------------*/
